Upgraded HRBF interface to be more Rustic + docs
Renamed HRBF to Hrbf in all type occurances to comply with Rust API
guidelines.

Built a builder interface for building HRBFs. This way it is impossible
to create an invalid/empty HRBF field. The intent is now better
represented in the interface.

The above changes also force the bundled C bindings to change slightly:
building an HRBF potential or the corresponding system is now done with
on C API call instead of two.

The C API now uses enums instead of ints to select kernel types which is
more idomatic in both C and Rust.

All dependencies were upgraded to latest versions.

Crate level documentation was added with a basic representative example.
Docs were also formatted better.

A warn(missing_docs) was added to lib.rs to keep track that
documentation is complete.

Bumped crate to v0.3.0.